refactor(peer): split atomic ingress boundaries
