+++ crates/atm-core/src/send/mod.rs
@@ -1,6 +1,5 @@
 //! Send command service implementation and post-send hook handling.
 
-use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 use std::time::Duration;
 
@@ -45,6 +44,7 @@ pub(crate) mod hook;
 pub mod input;
 #[doc(hidden)]
 pub(crate) mod nudge_template;
+mod peer_array;
 mod persistence;
 mod post_write;
 mod recipient;
@@ -509,9 +509,10 @@ pub fn prepare_peer_writes_atomically(
     observability: &dyn ObservabilityPort,
     runtime: &LocalServiceRuntime,
 ) -> Result<Vec<PreparedWrite>, AtmError> {
-    prepare_peer_writes_atomically_with_runtime(requests, observability, runtime)
+    peer_array::prepare_with_runtime(requests, observability, runtime)
 }
 
+#[cfg(test)]
 fn prepare_peer_writes_atomically_with_runtime<
     R: RetainedServiceRuntime + RetainedMailboxRuntime + crate::boundary::sealed::Sealed,
 >(
@@ -519,192 +520,7 @@ fn prepare_peer_writes_atomically_with_runtime<
     observability: &dyn ObservabilityPort,
     runtime: &R,
 ) -> Result<Vec<PreparedWrite>, AtmError> {
-    if requests.is_empty() {
-        return Err(AtmError::validation(
-            "peer message array must contain at least one write",
-        ));
-    }
-
-    let mut seen_origin_ids = HashSet::with_capacity(requests.len());
-    let mut pending = Vec::with_capacity(requests.len());
-    let mut records = Vec::with_capacity(requests.len());
-
-    for request in requests {
-        validate_write_provenance(
-            WriteIngress::Canonical,
-            WriteProvenance {
-                target_host: request.to.as_ref().and_then(|address| address.host()),
-                authenticated_source_host: request.authenticated_source_host.as_ref(),
-                origin_message_id: request.origin_message_id.is_some(),
-                origin_timestamp: request.origin_timestamp.is_some(),
-            },
-        )?;
-        if request.acknowledges_message_id.is_some() {
-            return Err(AtmError::validation(
-                "peer message arrays cannot contain acknowledgement writes",
-            ));
-        }
-        if request.dry_run {
-            return Err(AtmError::validation(
-                "peer message arrays cannot contain dry-run writes",
-            ));
-        }
-        let message_id = request.origin_message_id.ok_or_else(|| {
-            AtmError::validation("authenticated peer write is missing an origin message ID")
-        })?;
-        if !seen_origin_ids.insert(message_id) {
-            return Err(AtmError::validation(format!(
-                "peer message array contains duplicate origin message ID {message_id}"
-            )));
-        }
-
-        let context = prepare_send_context(runtime, &request)?;
-        let task_id = request.task_id.clone();
-        let requires_ack = request.requires_ack
-            || task_id.is_some()
-            || matches!(
-                &request.message_source,
-                SendMessageSource::File { path, .. } if file_policy::is_task_envelope(path)
-            );
-        let body = resolve_message_body(
-            &request.message_source,
-            &request.current_dir,
-            &request.home_dir,
-            &context.recipient.team,
-        )?;
-        let summary = summary::build_summary(&body, request.summary_override.clone());
-        let timestamp = request.origin_timestamp.unwrap_or_else(IsoTimestamp::now);
-        let mut envelope = build_send_envelope(
-            &request,
-            &context,
-            &body,
-            &summary,
-            message_id,
-            timestamp,
-            requires_ack,
-            task_id.clone(),
-        );
-        if envelope.parent_message_id.is_some() && envelope.thread_mode.is_some() {
-            let inbox_messages = persistence::load_store_backed_mailbox_projection(
-                runtime,
-                &request.home_dir,
-                &context.recipient.team,
-                &context.recipient.agent,
-            )?;
-            prepare_threaded_message(&mut envelope, &inbox_messages)?;
-        }
-        let record = boundary::Message {
-            team: context.recipient.team.clone(),
-            agent: context.recipient.agent.clone(),
-            message_key: boundary::MessageKey::from(message_id),
-            envelope: envelope.clone(),
-        };
-        let persistence = match runtime.load_message_record(
-            &request.home_dir,
-            &context.recipient.team,
-            &context.recipient.agent,
-            &record.message_key,
-        )? {
-            Some(existing) => match persistence::classify_existing_message(
-                existing,
-                &envelope,
-                message_id,
-                &context.recipient.team,
-                &context.recipient.agent,
-                request
-                    .authenticated_source_host
-                    .as_ref()
-                    .and_then(|source_host| {
-                        request
-                            .to
-                            .as_ref()
-                            .and_then(|destination| destination.host())
-                            .map(|destination_host| (source_host, destination_host))
-                    }),
-            )? {
-                DuplicateWriteDisposition::AlreadyDeliveredRemote => {
-                    DeliveryPersistenceResult::already_persisted(envelope)
-                }
-                DuplicateWriteDisposition::SameStorePeerReceipt => {
-                    DeliveryPersistenceResult::same_store_peer_receipt(envelope)
-                }
-                DuplicateWriteDisposition::NotDuplicate => {
-                    unreachable!("existing record is duplicate")
-                }
-            },
-            None => {
-                records.push(record);
-                DeliveryPersistenceResult::persisted(envelope)
-            }
-        };
-        pending.push((
-            request,
-            context,
-            body,
-            summary,
-            message_id,
-            timestamp,
-            requires_ack,
-            task_id,
-            persistence,
-        ));
-    }
-
-    // This is the only durable operation for new records in the accepted
-    // array. Validation and duplicate conflict checks above complete before it
-    // begins, so an error cannot leave a newly accepted subset behind.
-    runtime.persist_message_records_atomically(records)?;
-
-    pending
-        .into_iter()
-        .map(
-            |(
-                request,
-                context,
-                body,
-                summary,
-                message_id,
-                timestamp,
-                requires_ack,
-                task_id,
-                persistence,
-            )| {
-                let outcome = finalize_send_outcome(
-                    runtime,
-                    observability,
-                    &request,
-                    &context,
-                    &body,
-                    &summary,
-                    message_id,
-                    requires_ack,
-                    task_id,
-                    &persistence,
-                    DeliveryExecutionMode::Deferred,
-                )?;
-                Ok(PreparedWrite {
-                    outcome,
-                    outbound_request: request,
-                    persisted_timestamp: timestamp,
-                    post_write_needed: persistence.requires_post_write(),
-                    same_store_peer_receipt: persistence.duplicate_disposition
-                        == DuplicateWriteDisposition::SameStorePeerReceipt,
-                    #[cfg(test)]
-                    post_write: LocalPostWrite {
-                        post_send_config: context.post_send_config,
-                        recipient: context.recipient,
-                        delivery_snapshot: context.delivery_snapshot,
-                        messages: post_send_messages_from_persistence(
-                            &persistence,
-                            requires_ack,
-                            false,
-                        )?,
-                    },
-                    acknowledgement: None,
-                })
-            },
-        )
-        .collect()
+    peer_array::prepare_with_runtime(requests, observability, runtime)
 }
 
 /// The sole write pipeline. `acknowledges_message_id` selects only an
